rewrite triggers to only rebuild on visible rows
